touch: correct behaviors with relative times and --no-dereference
Corrects behavior to allow -r to be used together with -d. When -d
specifies an absolute date/time, the reference file is ignored;
when -d specifies a relative date/time (e.g., "+2 days"), the
relative value is applied to the times of the reference file.
To keep the code simple, this entailed extracting into a helper
function the existing logic that parses relative date/time
specifications and authoring some basic unit tests to pin the
behavior of that existing logic.

Also fixes some --no-dereference cases: now handles -h being used
with "-" as a file path, corrects some logic around when to use
std::fs::metadata and when to use std::fs::symlink_metadata.