uucore: add datetime parser
Add a relaxed datetime parser to uucore for use in `date` utility (and
potentially others). This datetime parser functions by using `chrono`s
own parsing utilities and a try/succeed approach to parsing. This commit
adds `date -d` functionality to the `date` utility by enabling parsing
of a number of date formats.

This implementation of the datetime parser has some drawbacks and some
positives. On the positive side:
- it was easy to implement
- it is easy to add more datetime formats to

In order to add additionally supported formats, a developer can add the
required format string to the `format` mod in `parse_datetime.rs`, and
then add it as a potential format to the relevant `fmts` vec.

On the negative:
- It is not easily customiseable beyond the supported `chrono` parsing
  formats. E.g., `chrono` does not currently support parsing offsets
without trailing zeros. `date -d "UTC+1"` should return a valid response
but `chrono` fails to parse this.
- Because it is an attempt driven parser, it is likely not that
  performant. I have not done any performance testing as part of this
change, but I would expect a custom parser to perform much better.