add windows version of copy_on_write for cp command
This is implemented using DeviceIoControl FSCTL_DUPLICATE_EXTENTS_TO_FILE,
and thus only works on ReFS (and possibly winbtrfs, I'm not sure).
While it should be possible to handle sparse modes on windows as well
this PR doesn't do that. Note that we do set the destination file as sparse
while copying it, this is because the target of FSCTL_DUPLICATE_EXTENTS_TO_FILE
must be at least as large as the source, and we don't want to write out zeros to the disk
before copying over the extents.