ptx: rewrite logic to support UTF8 input
The current logic uses offset addressing into the string converted as a
byte slice. This is problematic when the input is UTF8 encoded as the
offset in the byte slice doesn't necessarily correspond to a valid
character boundary in a utf8

Recode the logic to use String instead of a byte slice.

Signed-off-by: Guillaume Ranquet <granquet@baylibre.com>