touch: broaden no-op optimization to all mtimes that are 'now'
In isolation, this commit only creates a very minor optimization, at
the risk of a minor inaccuracy in case the caller guesses correctly
the precise time window during which determine_times() is run.
However, 'time_is_now' will later enable a required feature that is
explicitly tested in the GNU tests.